pyproject.toml
[project]
name = "plant-pred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "gdown>=5.2.0",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "streamlit>=1.49.1",
    "tensorflow>=2.20.0",
]
